:recycle: Break up validators.js into separate files.
Why? To make it easier to manage and maintain the validators.
Note that some validators are now backed by validate.js library.
