+++ web/src/utils/validators/is-email.js
@@ -1,0 +1,7 @@
+import { isEmail as isEmailValidator } from "validator"
+
+export const isEmail = (value) => {
+  return isEmailValidator(value) || "Invalid e-mail."
+}
+
+export default isEmail
