+++ web/src/utils/validators/is-integer.js
@@ -1,0 +1,4 @@
+export const isInteger = (v) =>
+  v == 0 || Number.isInteger(Number(v)) || "This field must be a number"
+
+export default isInteger
