+++ web/src/utils/validators/is-phone-number.js
@@ -1,0 +1,7 @@
+import { isMobilePhone } from "validator"
+
+export const isPhoneNumber = (value) => {
+  return isMobilePhone(value) || "Invalid phone number"
+}
+
+export default isPhoneNumber
